+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/equalshashcode/InputEqualsHashCodeSemanticTwo.java
@@ -1,0 +1,70 @@
+/*
+EqualsHashCode
+
+
+*/
+
+package com.puppycrawl.tools.checkstyle.checks.coding.equalshashcode;
+
+import java.awt.*;
+import java.io.ByteArrayOutputStream;
+import java.io.File;
+
+/**
+ * Test case for detecting simple semantic violations.
+ * @author Lars Kühne
+ **/
+class InputEqualsHashCodeSemanticTwo
+{
+    public class EqualsVsHashCode4
+    {
+        // in anon inner class
+        ByteArrayOutputStream bos1 = new ByteArrayOutputStream()
+        {
+            public boolean equals(Object a)
+            {
+                return true;
+            }
+
+            public int hashCode()
+            {
+                return 0;
+            }
+        };
+
+        ByteArrayOutputStream bos2 = new ByteArrayOutputStream()
+        {
+            public boolean equals(Object a) // violation 'without .* of 'hashCode()'.'
+            {
+                return true;
+            }
+        };
+    }
+
+    public void triggerEmptyBlockWithoutBlock()
+    {
+        // an if statement without a block to increase test coverage
+        if (true)
+            return;
+    }
+
+    // empty instance initializer
+    {
+    }
+
+    public class EqualsVsHashCode5
+    {
+        public <A> boolean equals(int a) // wrong arg type, ok even with generics
+        {
+            return a == 1;
+        }
+    }
+
+    public class EqualsVsHashCode6
+    {
+        public <A> boolean equals(Comparable<A> a)
+        {
+            return true;
+        }
+    }
+}
